requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
chromadb==0.5.5
click==8.1.3
colorlog==6.8.2
lxml==5.1.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.12.0
Pillow==10.0.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.0
typing_extensions==4.9.0
urllib3==2.0.5
